Merge updated JSON messages into Hive current table
This code adds the ability to specify command line options to the Scala code to identify the information needed to update a Hive table.  The messages that are to be updated are in JSON on HDFS and will use a unique key to identify order when dealing with the entire message history.